[xcompiler] Switch FusedEltwise CLAMP min/max attrs to I32Attr
XIR / xmodel emit CLAMP min/max as signless i32 attributes. The XCOMPILER
ODS-generated FusedEltwise op was emitting them as si64, causing
xir.qlinear_eltwise printed as `max = 1 : si64` from the ONNX path while
the xmodel path printed `max = 1 : i32`.

Add an `int32` mapping in gen_onnx_mlir.py for custom-op YAML attributes
so they lower to OptionalAttr<I32Attr>, mark min/max in the schema as
type: int32, regenerate XCOMPILEROps.td, and update the C++ consumers
(ReplaceQDQEltwisePass uses a new getI32Attr helper; FuseConvActivationPass
RELU6 detection compares against std::optional<int32_t>).

+++ src/Dialect/ONNX/Transforms/xmc/FuseConvActivationPass.cpp
@@ -150,8 +150,8 @@ static ActivationInfo getActivationInfo(Operation *op) {
     else if (opType == "LEAKYRELU")
       info.activationType = "LEAKYRELU";
     else if (opType == "CLAMP" &&
-             fusedOp.getMin() == std::optional<int64_t>(0) &&
-             fusedOp.getMax() == std::optional<int64_t>(6))
+             fusedOp.getMin() == std::optional<int32_t>(0) &&
+             fusedOp.getMax() == std::optional<int32_t>(6))
       info.activationType = "RELU6";
     else if (opType == "SIGMOID")
       info.activationType = "SIGMOID";
